+++ tests/test_complete_files.py
@@ -132,6 +132,13 @@ class TestCompleteFiles(unittest.TestCase):
         filename = "nodes_and_transform"
         create_test_from_filename(filename, self)
 
+    def test_attribute_texmode(self):
+        """Test per SVG object texmode with attribute"""
+        filename = "attribute_texmode"
+        create_test_from_filename(
+            filename, self, texmode="attribute", texmode_attribute="data-texmode"
+        )
+
 
 if __name__ == "__main__":
     unittest.main()